feat(evaluators): unify new-evaluator flow into a single sliding drawer
Replace the three stacked drawers (category selector → type selector →
evaluator editor) with a single drawer that slides between views, so the
flow no longer closes and re-opens on each step. Also adds a slide
transition to the evaluations wizard sidebar steps.
