+++ langwatch/src/components/evaluators/__tests__/EvaluatorCategorySelectorDrawer.test.tsx
@@ -8,7 +8,41 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { EvaluatorCategorySelectorDrawer } from "../EvaluatorCategorySelectorDrawer";
 
-// Mock dependencies
+// Stub the inline type-selector content so this test suite stays focused on
+// category-step behavior without pulling in tRPC-backed dependencies.
+vi.mock("../EvaluatorTypeSelectorContent", () => ({
+  categoryNames: {
+    expected_answer: "Expected Answer",
+    llm_judge: "LLM as Judge",
+    rag: "RAG Quality",
+    quality: "Quality Aspects",
+    safety: "Safety",
+  },
+  EvaluatorTypeSelectorContent: ({ category }: { category?: string }) => (
+    <div data-testid="mock-type-selector-content">
+      type-content-for-{category}
+    </div>
+  ),
+}));
+
+// The unified drawer now hosts the editor step in the same AnimatePresence as
+// the picker steps, so the shared controller/body/footer are imported at parent
+// level. Mock them here so the test environment doesn't need tRPC plumbing.
+vi.mock("../EvaluatorEditorShared", () => ({
+  useEvaluatorEditorController: () => ({
+    title: "Mock Editor",
+    hasUnsavedChanges: false,
+    onLocalConfigChange: undefined,
+    handleClose: vi.fn(),
+    handleSave: vi.fn(),
+    handleDiscard: vi.fn(),
+    handleApply: vi.fn(),
+  }),
+  EvaluatorEditorBody: () => <div data-testid="mock-editor-body" />,
+  EvaluatorEditorFooter: () => <div data-testid="mock-editor-footer" />,
+  EvaluatorEditorHeading: () => <span data-testid="mock-editor-heading" />,
+}));
+
 vi.mock("~/utils/compat/next-router", () => ({
   useRouter: () => ({
     push: vi.fn(),
@@ -32,7 +66,6 @@ vi.mock("~/hooks/useDrawer", () => ({
   getComplexProps: () => ({}),
 }));
 
-// Wrapper with Chakra provider
 const Wrapper = ({ children }: { children: React.ReactNode }) => (
   <ChakraProvider value={defaultSystem}>{children}</ChakraProvider>
 );
@@ -94,7 +127,7 @@ describe("EvaluatorCategorySelectorDrawer", () => {
   });
 
   describe("Navigation", () => {
-    it("opens evaluator type selector when selecting a category", async () => {
+    it("switches to the type view when selecting a category", async () => {
       const user = userEvent.setup();
       renderDrawer();
 
@@ -106,9 +139,17 @@ describe("EvaluatorCategorySelectorDrawer", () => {
         screen.getByTestId("evaluator-category-expected_answer"),
       );
 
-      expect(mockOpenDrawer).toHaveBeenCalledWith("evaluatorTypeSelector", {
-        category: "expected_answer",
+      await waitFor(() => {
+        expect(
+          screen.getByTestId("mock-type-selector-content"),
+        ).toHaveTextContent("type-content-for-expected_answer");
       });
+      // Old behavior delegated to a second drawer via openDrawer — the unified
+      // flow swaps views inline instead.
+      expect(mockOpenDrawer).not.toHaveBeenCalledWith(
+        "evaluatorTypeSelector",
+        expect.anything(),
+      );
     });
 
     it("calls onSelectCategory when selecting a category", async () => {
@@ -122,6 +163,22 @@ describe("EvaluatorCategorySelectorDrawer", () => {
       await user.click(screen.getByTestId("evaluator-category-llm_judge"));
 
       expect(mockOnSelectCategory).toHaveBeenCalledWith("llm_judge");
+    });
+
+    it("shows a back button in the type view that returns to the category list", async () => {
+      const user = userEvent.setup();
+      renderDrawer();
+
+      await user.click(screen.getByTestId("evaluator-category-safety"));
+
+      const backButton = await screen.findByTestId("back-button");
+      await user.click(backButton);
+
+      await waitFor(() => {
+        expect(
+          screen.getByText("Choose Evaluator Category"),
+        ).toBeInTheDocument();
+      });
     });
   });
