+++ langwatch/src/components/evaluators/__tests__/EvaluatorTypeSelectorDrawer.test.tsx
@@ -192,7 +192,25 @@ describe("EvaluatorTypeSelectorDrawer", () => {
   });
 
   describe("Navigation", () => {
-    it("opens evaluator editor when selecting an evaluator", async () => {
+    it("opens evaluator editor when selecting an evaluator (no onSelect provided)", async () => {
+      const user = userEvent.setup();
+      renderDrawer();
+
+      await waitFor(() => {
+        expect(screen.getByText("Exact Match")).toBeInTheDocument();
+      });
+
+      await user.click(screen.getByText("Exact Match"));
+
+      expect(mockOpenDrawer).toHaveBeenCalledWith(
+        "evaluatorEditor",
+        expect.objectContaining({
+          evaluatorType: expect.stringContaining("exact_match"),
+        }),
+      );
+    });
+
+    it("delegates to onSelect (and skips openDrawer) when a handler is provided", async () => {
       const user = userEvent.setup();
       renderDrawer({ onSelect: mockOnSelect });
 
@@ -202,12 +220,12 @@ describe("EvaluatorTypeSelectorDrawer", () => {
 
       await user.click(screen.getByText("Exact Match"));
 
-      // Should open evaluator editor drawer
-      expect(mockOpenDrawer).toHaveBeenCalledWith(
+      expect(mockOnSelect).toHaveBeenCalledWith(
+        expect.stringContaining("exact_match"),
+      );
+      expect(mockOpenDrawer).not.toHaveBeenCalledWith(
         "evaluatorEditor",
-        expect.objectContaining({
-          evaluatorType: expect.stringContaining("exact_match"),
-        }),
+        expect.anything(),
       );
     });
   });
